requirements.txt
anyio==3.6.2
autopep8==2.0.2
beautifulsoup4==4.12.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
Cython==0.29.34
Flask==2.2.3
Flask-Cors==3.0.10
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
numpy==1.24.2
oauthlib==3.2.2
pandas==1.5.3
pony==0.7.16
pycodestyle==2.10.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
requests==2.28.2
requests-oauthlib==1.3.1
rfc3986==1.5.0
selectolax==0.3.12
six==1.16.0
sniffio==1.3.0
soupsieve==2.4
tomli==2.0.1
tweepy==4.13.0
urllib3==1.26.15
Werkzeug==2.2.3
